# openapi/api.yaml
openapi: 3.0.3
info:
  title: Proof of Concept Spring Boot MySQL
  description: Proof of Concept Spring Boot MySQL with fields binary and JSON
  version: 1.0.0
  contact:
    name: Franclin
    email: f@francl.in
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /v1/intentions:
    post:
      summary: Create a new intention
      operationId: createIntention
      tags:
        - Intentions V1
      parameters:
        - in: header
          name: X-Correlation-ID
          description: Correlation ID
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Intention'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Intention'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        500:
          $ref: '#/components/responses/500'
    get:
      summary: Search intentions
      operationId: searchIntentions
      tags:
        - Intentions V1
      parameters:
        - in: header
          name: X-Correlation-ID
          description: Correlation ID
          schema:
            type: string
            format: uuid
          required: true
        - in: query
          name: page
          description: Page position
          schema:
            type: integer
            default: 0
          required: false
        - in: query
          name: size
          description: Size
          schema:
            type: integer
            default: 10
          required: false
        - in: query
          name: sortDirection
          description: Sort direction
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
          required: false
        - in: query
          name: sortField
          description: Sort field
          schema:
            type: string
            default: id
          required: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ResponseList'
                  - properties:
                      data:
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Intention'
  /v1/intentions/{id}:
    get:
      summary: Get an intention by ID
      operationId: getIntention
      tags:
        - Intentions V1
      parameters:
        - in: header
          name: X-Correlation-ID
          description: Correlation ID
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: id
          description: Intention ID
          schema:
            type: string
            format: uuid
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Intention'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
    patch:
      summary: Patch an intention by ID
      operationId: patchIntention
      tags:
        - Intentions V1
      parameters:
        - in: header
          name: X-Correlation-ID
          description: Correlation ID
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: id
          description: Intention ID
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Intention'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Intention'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
    put:
      summary: Update an intention by ID
      operationId: updateIntention
      tags:
        - Intentions V1
      parameters:
        - in: header
          name: X-Correlation-ID
          description: Correlation ID
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: id
          description: Intention ID
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Intention'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Intention'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
components:
  schemas:
    ResponseDataListPage:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        size:
          type: integer
    ResponseDataList:
      type: object
      properties:
        total:
          type: integer
        page:
          $ref: '#/components/schemas/ResponseDataListPage'
        items:
          type: array
          items:
            type: object
    ResponsePartialError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
    ResponsePartialData:
      type: object
      properties:
        data:
          type: object
    ResponseBase:
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string
    Response:
      allOf:
        - $ref: '#/components/schemas/ResponseBase'
        - $ref: '#/components/schemas/ResponsePartialData'
    ResponseError:
      allOf:
        - $ref: '#/components/schemas/ResponseBase'
        - $ref: '#/components/schemas/ResponsePartialError'
    ResponseList:
      allOf:
        - $ref: '#/components/schemas/Response'
        - properties:
            data:
              $ref: '#/components/schemas/ResponseDataList'
...
    Intention:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        contextId:
          type: string
          format: uuid
...
  responses:
    400:
      description: Bad Request
    401:
      description: Unauthorized
    403:
      description: Forbidden
    404:
      description: Not Found
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseError'